bitset: use uint64_t and add features
- Better efficiency by using larger type
- Additional macros/functions
- Convert uses for language mask

+++ vod/media_set.h
@@ -181,7 +181,7 @@ typedef struct {
 	uint32_t tracks_mask[MEDIA_TYPE_COUNT];
 	sequence_tracks_mask_t* sequence_tracks_mask;
 	sequence_tracks_mask_t* sequence_tracks_mask_end;
-	uint8_t* langs_mask;			// [LANG_MASK_SIZE]
+	size_t* langs_mask;			// [LANG_MASK_SIZE]
 	uint32_t version;
 	uint32_t width;
 	uint32_t height;
